requirements.txt
asgiref==3.4.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Cython==0.29.30
Django==3.2.13
django-cors-headers==3.10.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
glove==1.0.0
glove-python-binary==0.2.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.0
jsonparser==1.0
mysqlclient==2.0.1
nltk==3.6.7
numpy==1.19.5
pandas==1.1.5
pybind11==2.9.2
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
regex==2022.6.2
requests==2.27.1
scikit-learn==0.24.2
scikit-surprise==1.1.1
scipy==1.5.4
six==1.16.0
sqlparse==0.4.2
surprise==0.1
threadpoolctl==3.1.0
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
zipp==3.6.0
